+++ packages/core/src/interactivity/PersistentTooltipDataComponent.tsx
@@ -55,6 +55,7 @@ export const PersistentTooltipDataComponent = <
         ...props,
         key,
         style: componentStyle,
+        ...handlers,
         onClick: handleClick,
     })
 }