+++ packages/core/src/tooltips/types.ts
@@ -14,7 +14,10 @@ export type TooltipTitleProps = LegendTitleProps
 
 export type TooltipItemProps = Omit<LegendItemProps, 'interactive' | 'disabledStyle'>
 
-export type TooltipItemThemedProps = Omit<LegendItemThemedProps, 'interactive' | 'disabledStyle'>
+export type TooltipItemThemedProps = Omit<
+    LegendItemThemedProps,
+    'interactive' | 'disabledStyle' | 'allowDisableAll'
+>
 
 export interface TooltipProps
     extends Omit<
@@ -27,6 +30,7 @@ export interface TooltipProps
         | 'sizeTicks'
         | 'interactive'
         | 'disabledStyle'
+        | 'allowDisableAll'
     > {
     /** tooltip type */
     variant?: 'list'
@@ -41,7 +45,10 @@ export interface TooltipProps
 }
 
 export interface TooltipThemedProps
-    extends Omit<LegendThemedProps, 'interactive' | 'disabledStyle' | 'scaleSize'> {
+    extends Omit<
+        LegendThemedProps,
+        'interactive' | 'disabledStyle' | 'allowDisableAll' | 'scaleSize'
+    > {
     offset: NumericPositionSpec
     padding: FourSideSizeSpec
     anchor: AnchorSpec
@@ -59,14 +66,17 @@ export interface AxisTooltipThemedProps extends TooltipThemedProps {
     variant: SideVariant
 }
 
-export type TooltipItemListProps = Omit<LegendItemListProps, 'interactive' | 'disabledStyle'> & {
+export type TooltipItemListProps = Omit<
+    LegendItemListProps,
+    'interactive' | 'disabledStyle' | 'allowDisableAll'
+> & {
     /** array of ids */
     ids?: string[]
 }
 
 export type TooltipItemListThemedProps = Omit<
     LegendItemListThemedProps,
-    'interactive' | 'disabledStyle'
+    'interactive' | 'disabledStyle' | 'allowDisableAll'
 >
 
 /** Context */